Comment create/edit/update/delete

+++ database/migrations/2020_09_16_003955_add_multiple_column_to_posts.php
@@ -19,7 +19,7 @@ class AddMultipleColumnToPosts extends Migration
             $table->unsignedBigInteger('comments_count')->default(0);
             $table->unsignedBigInteger('shares_count')->default(0);
             $table->bigInteger('votes_count')->default(0);
-            $table->unsignedBigInteger('best_comment')->nullable();
+            $table->tinyInteger('best_comment')->nullable();
         });
     }
